requirements.txt
# newsletter_project/requirements.txt
python-dotenv
pydantic
requests
langchain-google-genai